+++ src/controllers/companyController.js
@@ -1,8 +1,14 @@
 import { companies } from "../models/index.js";
 import NotFoundError from "../errors/notFoundError.js";
 import validateCompany from "../usecases/validateCompany.js";
+import buildSearchQuery from "../usecases/buildSearchQuery.js";
 
 class CompanyController {
+  static async countCompanies() {
+    const companiesFound = await companies.find();
+    return companiesFound.length;
+  }
+  
   static async registerCompany(req, res, next) {
     try {
       validateCompany(req.body);
@@ -60,23 +66,6 @@ class CompanyController {
       next(err);
     }
   }
-
-  static async countCompanies() {
-    const companiesFound = await companies.find();
-    return companiesFound.length;
-  }
-}
-
-function buildSearchQuery(params) {
-  const { name, city, document } = params;
-
-  let search = {};
-
-  if (name) search.name = { $regex: name, $options: "i" };
-  if (city) search.city = city;
-  if (document) search.document = { $regex: document, $options: "i" };
-
-  return search;
 }
 
 export default CompanyController;
